Feature/creacion event model

+++ banking-account/cqrs-es/cqrs.core/src/main/java/com/banking/cqrs/core/events/BaseEvent.java
@@ -11,5 +11,5 @@ import lombok.experimental.SuperBuilder;
 @AllArgsConstructor
 @SuperBuilder
 public abstract class BaseEvent extends Message {
-    private int version;
+    private long version;
 }
